ajout slug; barre de recherche; protection route; admin, editeur, utilisateur;

+++ src/Controller/CategoriesController.php
@@ -5,31 +5,50 @@ namespace App\Controller;
 use App\Entity\Categories;
 use App\Form\CategoriesType;
 use App\Repository\CategoriesRepository;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
 /**
- * @Route("/categories")
+ * @Route("/admin/categories", name="categories_")
  */
 class CategoriesController extends AbstractController
 {
     /**
-     * @Route("/", name="categories_index", methods={"GET"})
+     * @IsGranted("ROLE_EDITOR")
+     * @Route("/", name="index", methods={"GET"})
      */
-    public function index(CategoriesRepository $categoriesRepository): Response
+    public function index(CategoriesRepository $categoriesRepository, PaginatorInterface $paginator, Request $request): Response
     {
+
+        $this->denyAccessUnlessGranted('ROLE_EDITOR');
+
+        $categories = $paginator->paginate(
+            //Selectionne toutes les données de la table "posts"
+            //getRepository attend en paramètre, l'entité avec laquelle on souhaite travailler
+            $categoriesRepository->findAll(),
+            //Le numero de la page, si aucun numero, on force la page 1
+            $request->query->getInt('page', 1),
+            //Nombre d'élément par page
+            10
+        );
+
         return $this->render('categories/index.html.twig', [
-            'categories' => $categoriesRepository->findAll(),
+            'categories' => $categories
         ]);
     }
 
     /**
-     * @Route("/new", name="categories_new", methods={"GET","POST"})
+     * @IsGranted("ROLE_EDITOR")
+     * @Route("/new", name="new", methods={"GET","POST"})
      */
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_EDITOR');
+
         $category = new Categories();
         $form = $this->createForm(CategoriesType::class, $category);
         $form->handleRequest($request);
@@ -39,8 +58,12 @@ class CategoriesController extends AbstractController
             $entityManager->persist($category);
             $entityManager->flush();
 
+            //Envoi d'un message de succès
+            $this->addFlash('success', 'La catégorie a bien été créée.');
+
             return $this->redirectToRoute('categories_index');
         }
+
 
         return $this->render('categories/new.html.twig', [
             'category' => $category,
@@ -49,25 +72,21 @@ class CategoriesController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="categories_show", methods={"GET"})
-     */
-    public function show(Categories $category): Response
-    {
-        return $this->render('categories/show.html.twig', [
-            'category' => $category,
-        ]);
-    }
-
-    /**
-     * @Route("/{id}/edit", name="categories_edit", methods={"GET","POST"})
+     * @IsGranted("ROLE_ADMIN")
+     * @Route("/edit/{id}", name="edit", methods={"GET","POST"})
      */
     public function edit(Request $request, Categories $category): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(CategoriesType::class, $category);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
+
+            //Envoi d'un message de succès
+            $this->addFlash('success', 'La catégorie a bien été modifiée.');
 
             return $this->redirectToRoute('categories_index');
         }
@@ -79,14 +98,20 @@ class CategoriesController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="categories_delete", methods={"DELETE"})
+     * @IsGranted("ROLE_ADMIN")
+     * @Route("/delete/{id}", name="delete", methods={"DELETE"})
      */
     public function delete(Request $request, Categories $category): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$category->getId(), $request->request->get('_token'))) {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+        
+        if ($this->isCsrfTokenValid('delete' . $category->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($category);
             $entityManager->flush();
+
+            //Envoi d'un message de succès
+            $this->addFlash('success', 'La catégorie a bien été supprimée.');
         }
 
         return $this->redirectToRoute('categories_index');
